backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
psutil
requests
python-dotenv
